Preserve Juju topology in provided jobs
In case of charms that need to proxy scrape relations on behalf of
other charms, we need to preserve the Juju topology labels built
into the scrape_config -> labels.

Since this use-case is somewhat of a corner-case for the Prometheus
charm library, it is not worth it to provide a first-level support
like allowing the charm using PrometheusConsumer to specify custom
scrape_metadata or the like.